+++ src/types/Props/typographyProps.ts
@@ -1,31 +1,18 @@
-import { ColorProps, FontSizeProps } from "@uiTypes";
-
-export type TextTransformOptions = "uppercase" | "capitalize" | "lowercase";
-export type TextTransformProps = {
-  textTransform?: TextTransformOptions;
-};
-
-export type TextAlignOptions = "left" | "center" | "right";
-export type TextAlignProps = {
-  textAlign?: TextAlignOptions;
-};
-
-export type TextDecorationOptions =
-  | "none"
-  | "dashed"
-  | "underline"
-  | "line-through";
-export type TextDecorationProps = {
-  textDecoration?: TextDecorationOptions;
-};
+import {
+  ColorProps,
+  FontSizeProps,
+  TextAlignProps,
+  TextDecorationProps,
+  TextTransformProps,
+} from "@uiTypes";
 
 export type TextAsProps = {
   as?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "p" | "span";
 };
 
-export type TextWithBackgroundClipProps = {};
 export type TypographyProps = {} & FontSizeProps &
   ColorProps &
-  TextAlignProps &
   TextAsProps &
-  TextTransformProps;
+  TextAlignProps &
+  TextTransformProps &
+  TextDecorationProps;